Scripts/requirements.txt
altgraph==0.17
backcall==0.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
DateTime==4.3
decorator==5.0.9
greenlet==1.0.0
idna==2.10
imgkit==1.2.2
ipython-genutils==0.2.0
jedi==0.18.0
kaleido==0.2.1
matplotlib-inline==0.1.2
mysql-connector==2.2.9
mysql-connector-python==8.0.23
numpy==1.20.2
pandas==1.2.3
parso==0.8.2
patsy==0.5.1
pexpect==4.8.0
pickleshare==0.7.5
plotly==4.14.3
plotly-express==0.4.1
prompt-toolkit==3.0.18
protobuf==3.15.7
psutil==5.8.0
ptyprocess==0.7.0
Pygments==2.9.0
pyinstaller==4.3
pyinstaller-hooks-contrib==2021.1
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.8.1
pyqtgraph==0.12.1
PyQtWebEngine==5.15.4
PyQtWebEngine-Qt5==5.15.2
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
retrying==1.3.3
scipy==1.6.2
six==1.15.0
soupsieve==2.2.1
SQLAlchemy==1.4.5
SQLAlchemy-Utils==0.36.8
statsmodels==0.12.2
traitlets==5.0.5
urllib3==1.26.4
wcwidth==0.2.5
zope.interface==5.3.0
